tantivy works!!
